test(picker-column-internal): migrate disabled
